+++ src/main/java/com/github/mnesikos/simplycats/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 ## [1.12.2-0.2.0] - 2022/07/19
 ### Added
 - Color inhibitor gene; adds silver tabbies and smokes!
+- Blue-eyed and red-eyed albino alleles; true albino cats!
 
 ## [1.12.2-0.0.4.1] - 2021/08/25
 ### Fixed